--- src/main/java/org/rest/book/ms/model/Employee.java
@@ -1,4 +1,0 @@
-package org.rest.book.ms.model;
-
-public record Employee(int id, String name) {
-}
