--- src/main/java/org/rest/book/ms/port/NumberPort.java
@@ -1,17 +1,0 @@
-package org.rest.book.ms.port;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
-import org.rest.book.ms.model.IsbnNumber;
-
-@RegisterRestClient(configKey = "number-ms")
-public interface NumberPort {
-
-    @GET
-    @Path("/v1/numbers")
-    @Produces(MediaType.APPLICATION_JSON)
-    IsbnNumber getIsbnNumber();
-}
