--- src/main/java/org/rest/book/ms/util/BookConstants.java
@@ -1,5 +1,0 @@
-package org.rest.book.ms.util;
-
-public class BookConstants {
-    public static final String FAULT_TOLERANT_GRAPHQL_RESPONSE = "Hello Stranger";
-}
