+++ src/main/java/org/rest/resource/ms/model/Employee.java
@@ -1,0 +1,4 @@
+package org.rest.resource.ms.model;
+
+public record Employee(int id, String name) {
+}
