+++ src/main/java/org/rest/resource/ms/util/BookConstants.java
@@ -1,0 +1,5 @@
+package org.rest.resource.ms.util;
+
+public class BookConstants {
+    public static final String FAULT_TOLERANT_GRAPHQL_RESPONSE = "Hello Stranger";
+}
