--- src/test/java/org/rest/book/ms/IT/GreetResourceIT.java
@@ -1,9 +1,0 @@
-package org.rest.book.ms.IT;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-import org.rest.book.ms.resource.GreetResourceTest;
-
-@QuarkusIntegrationTest
-class GreetResourceIT extends GreetResourceTest {
-    // Execute the same tests but in packaged mode.
-}
