--- src/test/java/org/rest/book/ms/resource/BookResourceTest.java
@@ -1,36 +1,0 @@
-package org.rest.book.ms.resource;
-
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.*;
-
-@QuarkusTest
-public class BookResourceTest {
-
-
-    @Test
-    void should_create_a_book_when_data_is_passed() {
-        var title = "The Lord of the Rings";
-        var author = "J.R.R. Tolkien";
-        var genre = "Fantasy";
-        var yearOfPublication = 1954;
-        given()
-                .formParam("title", title)
-                .formParam("author", author)
-                .formParam("yearOfPublication", yearOfPublication)
-                .formParam("genre", genre)
-                .when()
-                .post("/v1/save")
-                .then()
-                .statusCode(201)
-                .body("author", is(author))
-                .body("isbn_13", is("13-1234567890123"))
-                .body("year_of_publication", is(yearOfPublication))
-                .body("genre", is(genre))
-                .body("creation_date", notNullValue())
-                .body(not(hasKey("yearOfPublication")))
-                .body(not(hasKey("creationDate")));
-    }
-}
