+++ src/test/java/org/rest/resource/ms/IT/ArtistResourceIT.java
@@ -1,0 +1,9 @@
+package org.rest.resource.ms.IT;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+import org.rest.resource.ms.resource.ArtistResourceTest;
+
+@QuarkusIntegrationTest
+class ArtistResourceIT extends ArtistResourceTest {
+    // Execute the same tests but in packaged mode.
+}
